+++ src/main/java/at/hannibal2/skyhanni/features/garden/GardenAPI.kt
@@ -5,6 +5,7 @@ import at.hannibal2.skyhanni.data.IslandType
 import at.hannibal2.skyhanni.data.ScoreboardData
 import at.hannibal2.skyhanni.events.*
 import at.hannibal2.skyhanni.features.garden.composter.ComposterOverlay
+import at.hannibal2.skyhanni.features.garden.contest.FarmingContestAPI
 import at.hannibal2.skyhanni.features.garden.farming.GardenBestCropTime
 import at.hannibal2.skyhanni.features.garden.inventory.SkyMartCopperPrice
 import at.hannibal2.skyhanni.utils.ItemUtils.getInternalName
@@ -161,7 +162,8 @@ object GardenAPI {
 
     fun isSpeedDataEmpty() = cropsPerSecond.values.sum() < 0
 
-    fun hideExtraGuis() = ComposterOverlay.inInventory || AnitaMedalProfit.inInventory || SkyMartCopperPrice.inInventory
+    fun hideExtraGuis() = ComposterOverlay.inInventory || AnitaMedalProfit.inInventory ||
+                SkyMartCopperPrice.inInventory || FarmingContestAPI.inInventory
 
     fun clearCropSpeed() {
         for (type in CropType.values()) {